+++ apps/design-system/content/docs/fragments/page-header.mdx
@@ -21,4 +21,4 @@ fragment: true
 - `PageHeaderTitle` - Primary page heading (h1)
 - `PageHeaderDescription` - Supporting text below title
 - `PageHeaderAside` - Container for action buttons (should be inside PageHeaderMeta, has `shrink-0`)
-- `PageHeaderFooter` - Container for tab navigation (NavMenu, should be last child)
+- `PageHeaderNavigationTabs` - Container for tab navigation (NavMenu, should be last child)
